updating Dockerfile
1. developing AWS services for cost effective production.
2. updated Dockerfile blueprint to have maven in Docker Image
3. updated POM.xml file to run test cases when building maven.
