+++ pom.xml
@@ -17,6 +17,9 @@
 	<description>Utopia Auth MS</description>
 	<properties>
 		<java.version>1.8</java.version>
+		  <sonar.projectKey>turalss_AuthMS</sonar.projectKey>
+  <sonar.organization>turalss</sonar.organization>
+  <sonar.host.url>https://sonarcloud.io</sonar.host.url>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -92,22 +95,54 @@
 			<artifactId>spring-boot-freemarker-email</artifactId>
 			<version>0.6.3</version>
 		</dependency>
+				<dependency>
+			<groupId>org.springdoc</groupId>
+			<artifactId>springdoc-openapi-ui</artifactId>
+			<version>1.5.2</version>
+		</dependency>
+		
+	
+   <dependency>
+     <groupId>org.hrorm</groupId>
+     <artifactId>hrorm</artifactId>
+     <version>0.1.0-rc2</version>
+   </dependency>
 	</dependencies>
-
+	
 	<build>
 		<plugins>
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
-			<plugin>
-  <artifactId>maven-surefire-plugin</artifactId>
-  <version>2.10</version>
-  <configuration>
-    <useFile>false</useFile>
-  </configuration>
-</plugin>
 		</plugins>
 	</build>
+<profiles>
+		<profile>
+			<id>production</id>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>com.vaadin</groupId>
+						<artifactId>vaadin-maven-plugin</artifactId>
+						<version>${vaadin.version}</version>
+						<executions>
+							<execution>
+								<id>frontend</id>
+								<phase>compile</phase>
+								<goals>
+									<goal>prepare-frontend</goal>
+									<goal>build-frontend</goal>
+								</goals>
+								<configuration>
+									<productionMode>true</productionMode>
+								</configuration>
+							</execution>
+						</executions>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+	</profiles>
 
 </project>
