+++ src/main/resources/application.properties
@@ -3,9 +3,9 @@ server.port=8083
 logging.level.org.springframework.web=debug
 
 spring.mvc.view.prefix=/WEB-INF/
-spring.datasource.url=${DB_URL}
-spring.datasource.username=${DB_USERNAME}
-spring.datasource.password=${DB_PASSWORD}
+spring.datasource.url=jdbc:mysql://utopiair.c7huwkid5m6t.us-east-1.rds.amazonaws.com:3306/utopia
+spring.datasource.username=root
+spring.datasource.password=rootroot
 spring.jpa.hibernate.ddl-auto=update
 
 spring.mail.default-encoding=UTF-8
