--- src/test/java/com/utopia/auth/UtopiaauthmsApplicationTests.java
@@ -1,9 +1,0 @@
-package com.utopia.auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UtopiaauthmsApplicationTests {
-
-}
